Moved PlaceholderDefaults to ptoper package and added docs
