+++ evervault-inputs/src/main/java/com/evervault/sdk/input/ui/RowsPaymentCardInput.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import com.evervault.sdk.input.defaults.PlaceholderDefaults
+import com.evervault.sdk.input.defaults.placeholder.PlaceholderDefaults
 import com.evervault.sdk.input.defaults.placeholder.model.PlaceholderTexts
 
 @Composable
