+++ evervault-inputs/src/main/java/com/evervault/sdk/input/ui/card/InlinePaymentCard.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
-import com.evervault.sdk.input.defaults.PlaceholderDefaults
+import com.evervault.sdk.input.defaults.placeholder.PlaceholderDefaults
 import com.evervault.sdk.input.defaults.placeholder.model.PlaceholderTexts
 import com.evervault.sdk.input.model.card.PaymentCardData
